backend/requirements.txt
# Django backend dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
joblib==1.3.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
